Add basic ledger
Create a state type of ledger. A map from account hash aka pubkey to its
amount for the given chain. Initialization happens deterministically.

Frankly speaking, a state of accounts doesn't mean ledger. However, it's
a great prototype of ledger and later on it will probably be changed to
somthing else.

As for now, we can experiment with transactions using such a thing.