Refactor Tx structure
Create SignedTx struct which contains an encrypted tx and author key to
validate it. The reason behind such a change is that we have
that information inside a block anyway and having this structure on
its own is gonna help for communication between
a mining instance and a wallet which proposes transactions.

Make getTx a method of Block. I dunno why to be honest. It might look
better though.